fix wheel crawler service and other improvements
