+++ nix/python.nix
@@ -1,19 +1,15 @@
 let
   mach-nix = import (builtins.fetchGit {
     url = "https://github.com/DavHau/mach-nix/";
-    ref = "master";
-    rev = "fa2bb2d33fb9b9dc3113046e4fcc16088f56981a";
+    ref = "2.0.0";
   });
 in
 mach-nix.mkPython {
   requirements = ''
     requests
-    psycopg2
+    psycopg2 >= 2.8.0
     pkginfo
     peewee
     bounded-pool-executor
-    httpio
-    hanging_threads
-    scrapy
   '';
 }
