+++ src/crawl_sdist_deps.py
@@ -135,17 +135,16 @@ def extract_requirements(job: PackageJob):
 
 def get_jobs(pypi_fetcher_dir, bucket, processed, amount=1000):
     pypi_dict = LazyBucketDict(f"{pypi_fetcher_dir}/pypi", restrict_to_bucket=bucket)
-    total_count = 0
     jobs = []
     names = list(pypi_dict.by_bucket(bucket).keys())
-    shuffle(names)
+    total_nr = 0
     for pkg_name in names:
         for ver, release_types in pypi_dict[pkg_name].items():
             if 'sdist' not in release_types:
                 continue
             if (pkg_name, ver) in processed:
                 continue
-            total_count += 1
+            total_nr += 1
             release = release_types['sdist']
             if len(jobs) <= amount:
                 jobs.append(PackageJob(
@@ -153,9 +152,12 @@ def get_jobs(pypi_fetcher_dir, bucket, processed, amount=1000):
                     ver,
                     f"https://files.pythonhosted.org/packages/source/{pkg_name[0]}/{pkg_name}/{release[1]}",
                     release[0],
-                    total_count,
+                    0,
                 ))
-    print(f"Bucket {bucket}: Planning execution of {len(jobs)} jobs out of {total_count} total jobs for this bucket")
+    shuffle(jobs)
+    for i, job in enumerate(jobs):
+        job.idx = i
+    print(f"Bucket {bucket}: Planning execution of {len(jobs)} jobs out of {total_nr} total jobs for this bucket")
     return jobs
